% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossTable.R
\name{crossTable}
\alias{crossTable}
\alias{crossTable,itemMatrix-method}
\title{Cross-tabulate joint occurrences across pairs of items}
\usage{
crossTable(x, ...)

\S4method{crossTable}{itemMatrix}(
  x,
  measure = c("count", "support", "probability", "lift"),
  sort = FALSE
)
}
\arguments{
\item{x}{object to be cross-tabulated (\link{transactions} or
\link{itemMatrix}).}

\item{...}{additional arguments.}

\item{measure}{measure to return. Default is co-occurrence counts.}

\item{sort}{sort the items by support.}
}
\value{
A symmetric matrix of n x n, where n is the number of items times
in \code{x}. The matrix contains the co-occurrence counts between pairs of
items.
}
\description{
Provides the generic function \code{crossTable()} and a method to
cross-tabulate joint occurrences across all pairs of items.
}
\examples{
data("Groceries")

ct <- crossTable(Groceries, sort = TRUE)
ct[1:5, 1:5]

sp <- crossTable(Groceries, measure = "support", sort = TRUE)
sp[1:5, 1:5]

lift <- crossTable(Groceries, measure = "lift", sort = TRUE)
lift[1:5, 1:5]

}
\seealso{
Other itemMatrix and transactions functions:
\code{\link[=abbreviate]{abbreviate()}},
\code{\link{c}},
\code{\link[=duplicated]{duplicated()}},
\code{\link{extract}},
\code{\link{hierarchy}},
\code{\link{image}},
\code{\link[=inspect]{inspect()}},
\code{\link[=is.superset]{is.superset()}},
\code{\link[=itemFrequency]{itemFrequency()}},
\code{\link[=itemFrequencyPlot]{itemFrequencyPlot()}},
\code{\link{itemMatrix-class}},
\code{\link{itemwiseSetOps}},
\code{\link[=match]{match()}},
\code{\link[=merge]{merge()}},
\code{\link[=random.transactions]{random.transactions()}},
\code{\link[=sample]{sample()}},
\code{\link{sets}},
\code{\link[=size]{size()}},
\code{\link[=supportingTransactions]{supportingTransactions()}},
\code{\link{tidLists-class}},
\code{\link{transactions-class}},
\code{\link[=unique]{unique()}}
}
\author{
Michael Hahsler
}
\concept{itemMatrix and transactions functions}
\keyword{models}
