% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Groceries}
\alias{Groceries}
\title{Groceries Data Set}
\format{
Object of class \link{transactions}.
}
\source{
The data set is provided for arules by Michael Hahsler, Kurt Hornik
and Thomas Reutterer.
}
\description{
The \code{Groceries} data set contains 1 month (30 days) of real-world
point-of-sale transaction data from a typical local grocery outlet.  The
data set contains 9835 transactions and the items are aggregated to 169
categories.
}
\details{
If you use this data set in your paper, please refer to the paper in the
references section.
}
\references{
Michael Hahsler, Kurt Hornik, and Thomas Reutterer (2006)
Implications of probabilistic data modeling for mining association rules. In
M. Spiliopoulou, R. Kruse, C. Borgelt, A. Nuernberger, and W. Gaul, editors,
\emph{From Data and Information Analysis to Knowledge Engineering, Studies
in Classification, Data Analysis, and Knowledge Organization}, pages
598--605. Springer-Verlag.
}
\author{
Michael Hahsler
}
\keyword{datasets}
