% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.significant.R
\name{is.significant}
\alias{is.significant}
\alias{is.significant,rules-method}
\title{Find Significant Rules}
\usage{
is.significant(x, ...)

\S4method{is.significant}{rules}(
  x,
  transactions = NULL,
  method = "fisher",
  alpha = 0.01,
  adjust = "none",
  reuse = TRUE,
  ...
)
}
\arguments{
\item{x}{a set of \link{rules}.}

\item{...}{further arguments are passed on to \code{\link[=interestMeasure]{interestMeasure()}}.}

\item{transactions}{optional set of \link{transactions}. Only needed if not sufficient
interest measures are available in \code{x}. If the test should be performed
on a transaction set different then the one used for mining (use \code{reuse = FALSE}).}

\item{method}{test to use. Options are \code{"fisher"}, \code{"chisq"}. Note that
the contingency table is likely to have cells with low expected values and
that thus Fisher's Exact Test might be more appropriate than the chi-squared
test.}

\item{alpha}{required significance level.}

\item{adjust}{method to adjust for multiple comparisons. Some options are
\code{"none"}, \code{"bonferroni"}, \code{"holm"}, \code{"fdr"}, etc. (see
\code{\link[stats:p.adjust]{stats::p.adjust()}} for more methods)}

\item{reuse}{logical indicating if information in the quality slot should be
reuse for calculating the measures.}
}
\value{
returns a logical vector indicating which rules are significant.
}
\description{
Provides the generic functions \code{is.significant()} and the method to
find significant \link{rules}.
}
\details{
The implementation for association rules uses Fisher's exact test with
correction for multiple comparisons to test the null hypothesis that the LHS
and the RHS of the rule are independent.  Significant rules have a p-value
less then the specified significance level alpha (the null hypothesis of
independence is rejected). See Hahsler and Hornik (2007) for details.
}
\examples{
data("Income")
rules <- apriori(Income, support = 0.2)
is.significant(rules)

rules[is.significant(rules)]

# Adjust P-values for multiple comparisons
rules[is.significant(rules, adjust = "bonferroni")]
}
\references{
Hahsler, Michael and Kurt Hornik (2007). New probabilistic
interest measures for association rules. \emph{Intelligent Data Analysis},
11(5):437--455.
\doi{10.3233/IDA-2007-11502}
}
\seealso{
\code{\link[stats:p.adjust]{stats::p.adjust()}}

Other interest measures:
\code{\link{confint}},
\code{\link[=coverage]{coverage()}},
\code{\link[=interestMeasure]{interestMeasure()}},
\code{\link[=is.redundant]{is.redundant()}},
\code{\link[=support]{support()}}

Other postprocessing:
\code{\link[=is.closed]{is.closed()}},
\code{\link[=is.generator]{is.generator()}},
\code{\link[=is.maximal]{is.maximal()}},
\code{\link[=is.redundant]{is.redundant()}},
\code{\link[=is.superset]{is.superset()}},
\code{\link[=ruleInduction]{ruleInduction()}}

Other associations functions:
\code{\link[=abbreviate]{abbreviate()}},
\code{\link{associations-class}},
\code{\link{c}},
\code{\link[=duplicated]{duplicated()}},
\code{\link{extract}},
\code{\link[=inspect]{inspect()}},
\code{\link[=is.closed]{is.closed()}},
\code{\link[=is.generator]{is.generator()}},
\code{\link[=is.maximal]{is.maximal()}},
\code{\link[=is.redundant]{is.redundant()}},
\code{\link[=is.superset]{is.superset()}},
\code{\link{itemsets-class}},
\code{\link[=match]{match()}},
\code{\link{rules-class}},
\code{\link[=sample]{sample()}},
\code{\link{sets}},
\code{\link[=size]{size()}},
\code{\link[=sort]{sort()}},
\code{\link[=unique]{unique()}}
}
\author{
Michael Hahsler
}
\concept{associations functions}
\concept{interest measures}
\concept{postprocessing}
\keyword{manip}
