% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_arules-package.R
\docType{package}
\name{arules-package}
\alias{arules}
\alias{arules-package}
\title{arules: Mining Association Rules and Frequent Itemsets}
\description{
Provides the infrastructure for representing, manipulating and analyzing transaction data and patterns (frequent itemsets and association rules). Also provides C implementations of the association mining algorithms Apriori and Eclat. Hahsler, Gruen and Hornik (2005) \doi{10.18637/jss.v014.i15}.
}
\section{Key functions and Classes}{

\itemize{
\item Mining Algorithms: \code{\link[=apriori]{apriori()}}, \code{\link[=eclat]{eclat()}}
\item Classes: \link{transactions}, \link{rules}, \link{itemsets}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mhahsler/arules}
  \item Report bugs at \url{https://github.com/mhahsler/arules/issues}
}

}
\author{
\strong{Maintainer}: Michael Hahsler \email{mhahsler@lyle.smu.edu} (\href{https://orcid.org/0000-0003-2716-1405}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Christian Buchta [copyright holder]
  \item Bettina Gruen [copyright holder]
  \item Kurt Hornik (\href{https://orcid.org/0000-0003-4198-9911}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Christian Borgelt [contributor, copyright holder]
  \item Ian Johnson [contributor]
  \item Makhlouf Ledmi [contributor]
}

}
\keyword{internal}
