% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Epub.R
\docType{data}
\name{Epub}
\alias{Epub}
\title{The Epub Transactions Data Set}
\format{
Object of class \link{transactions}
with 15729 transactions
and 936 items.
Item labels are document IDs of the form \code{"doc\\_11d"}.
Session IDs and time stamps for transactions are also provided as transaction information.
}
\source{
Provided by Michael Hahsler from the custom information system ePub-WU
at \url{https://epub.wu-wien.ac.at} (which is now replaced by eprint).
}
\description{
The \code{Epub} data set contains the download history of documents from the
electronic publication platform of the Vienna University of Economics and
Business Administration.  The data was recorded between Jan 2003 and Dec
2008.
}
\examples{
 
data(Epub)
inspect(head(Epub))
}
\author{
Michael Hahsler
}
\keyword{datasets}
