\name{Epub}
\alias{Epub}
\docType{data}
\title{Epub Data Set}
\description{
  The \code{Epub} data set contains the download history of documents
  from the electronic publication platform of the Vienna University of
  Economics and Business Administration.
  The data was recorded between Jan 2003 and Dec 2008.
}
\usage{data(Epub)}
\format{
  Object of class \code{\linkS4class{transactions}} with
  15729 transactions and 936 items.
  Item labels are document IDs of the from \code{"doc\_11d"}.
  Session IDs and time stamps for transactions are also provided.
}
\source{
  Provided by Michael Hahsler from ePub-WU at
  \url{http://epub.wu-wien.ac.at}.
}
\author{Michael Hahsler}
\keyword{datasets}
