% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Epub}
\alias{Epub}
\title{Epub Data Set}
\format{
Object of class \code{\linkS4class{transactions}} with 15729
transactions and 936 items.  Item labels are document IDs of the from
\code{"doc\_11d"}.  Session IDs and time stamps for transactions are also
provided.
}
\source{
Provided by Michael Hahsler from ePub-WU at
\url{https://epub.wu-wien.ac.at}.
}
\description{
The \code{Epub} data set contains the download history of documents from the
electronic publication platform of the Vienna University of Economics and
Business Administration.  The data was recorded between Jan 2003 and Dec
2008.
}
\author{
Michael Hahsler
}
\keyword{datasets}
