% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\alias{"[-methods"}
\alias{[,itemMatrix,ANY,ANY,ANY-method}
\alias{[,transactions,ANY,ANY,ANY-method}
\alias{[,tidLists,ANY,ANY,ANY-method}
\alias{[,rules,ANY,ANY,ANY-method}
\alias{[,itemsets,ANY,ANY,ANY-method}
\title{Methods for "[": Extraction or Subsetting in Package 'arules'}
\usage{
\S4method{[}{itemMatrix,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{transactions,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{tidLists,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{rules,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{itemsets,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{an object of class \link{itemMatrix}, \link{transactions} or \link{associations}.}

\item{i}{select rows/sets.}

\item{j}{select columns/items.}

\item{...}{further arguments are ignored.}

\item{drop}{ignored.}
}
\description{
Methods for \code{"["}, i.e., extraction or subsetting in package
\pkg{arules}. Subsetting can be done by integers containing column/row
numbers, vectors of logicals or strings containing parts of item labels.
}
\section{Methods}{
 \describe{ \item{[}{\code{signature(x = "itemMatrix", i =
"ANY", j = "ANY", drop= "ANY")}; extracts parts of an \code{itemMatrix}.
The first argument selects rows (e.g., transactions or rules) and the second
argument selects columns (items).  Either argument can be omitted to select
all rows or columns.} \item{[}{\code{signature(x = "itemsets", i = "ANY", j
= "ANY", drop= "ANY")}; extracts a subset of itemsets and the associated
quality measures.  \code{j} has to be missing. } \item{[}{\code{signature(x
= "rules", i = "ANY", j = "ANY", drop= "ANY")}; extracts a subset of rules
and the associated quality measures.  \code{j} has to be missing. }
\item{[}{\code{signature(x = "transactions", i = "ANY", j = "ANY", drop=
"ANY")}; extracts a subset of transactions/items from a transactions object
(a binary incidence matrix).  \code{i} and \code{j} can be numeric where
\code{i} selects transactions and \code{j} selects items. }
\item{[}{\code{signature(x = "tidLists", i = "ANY", j = "ANY", drop=
"ANY")}; extracts parts (transaction ID vectors) from \code{tidLists}.
\code{i} selects the items or itemsets and \code{j} selects transactions in
the lists.} }
}

\examples{
data(Adult)
Adult

## select first 10 transactions
Adult[1:10]

## select first 10 items for first 100 transactions
Adult[1:100, 1:10]

## select the first 100 transactions for the items containing
## "income" or "age=Young" in their labels
Adult[1:100, c("income=small", "income=large" ,"age=Young")]
}
\seealso{
Other associations functions: 
\code{\link{abbreviate}()},
\code{\link{associations-class}},
\code{\link{c}()},
\code{\link{duplicated}()},
\code{\link{inspect}()},
\code{\link{is.closed}()},
\code{\link{is.generator}()},
\code{\link{is.maximal}()},
\code{\link{is.significant}()},
\code{\link{is.superset}()},
\code{\link{match}()},
\code{\link{rules-class}},
\code{\link{sample}()},
\code{\link{sets}},
\code{\link{size}()},
\code{\link{sort}()},
\code{\link{unique}()}

Other itemMatrix and transactions functions: 
\code{\link{abbreviate}()},
\code{\link{crossTable}()},
\code{\link{c}()},
\code{\link{duplicated}()},
\code{\link{hierarchy}},
\code{\link{image}()},
\code{\link{inspect}()},
\code{\link{is.superset}()},
\code{\link{itemFrequencyPlot}()},
\code{\link{itemFrequency}()},
\code{\link{itemMatrix-class}},
\code{\link{match}()},
\code{\link{merge}()},
\code{\link{random.transactions}()},
\code{\link{sample}()},
\code{\link{sets}},
\code{\link{size}()},
\code{\link{supportingTransactions}()},
\code{\link{tidLists-class}},
\code{\link{transactions-class}},
\code{\link{unique}()}
}
\author{
Michael Hahsler
}
\concept{associations functions}
\concept{itemMatrix and transactions functions}
\keyword{array}
