% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fim4r.R
\name{fim4r}
\alias{fim4r}
\title{Interface to Mining Algorithms from fim4r}
\usage{
fim4r(transactions, method = NULL, target = "frequent", ...)
}
\arguments{
\item{transactions}{a \link{transactions} object}

\item{method}{the algorithm to be used. One of: "apriori", "eclat", "fpgrowth",
"carpenter" (closed itemsets only), "ista", "relim", "sam"}

\item{target}{the target type. One of: "frequent",
"closed", "maximal", "generators" or "rules".}

\item{...}{further arguments are passed on to the \code{fim4r.x()} in
package \pkg{fim4r} (\code{x} is the chosen method). Minimum support and
minimum confidence can be set as parameters \code{supp} and \code{conf}
(note the range is \eqn{[0, 100]} percent).}
}
\value{
An object of class \link{itemsets} or \link{rules}.
}
\description{
Interfaces the algorithms implemented in fim4r. The algorithms include:
apriori, arules, eclat, fpgrowth, carpenter, estpsp, fim, genpsp, ista,
relim, sam.
}
\details{
Note that the package \pkg{fim4r} is not available via CRAN. If needed,
the \code{fim4r()} function installs the package automatically.

Note on additional parameters for \code{fim4r.x} functions:
\itemize{
\item \code{report} is used internally and cannot be specified.
\item method \code{"carpenter"} can only mine closed itemsets.
\item for \code{appear}, the item ID needs to be used and not the item label.
}
}
\examples{
\dontrun{
data(Adult)

# note that fim4r specifies support and confidence out of 100\%
r <- fim4r(Adult, method = "fpgrowth", target = "rules", supp = 50, conf = 80)
r

inspect(head(r, by = "lift"))
}
}
\references{
Christian Borgelt, fimi4r: Frequent Item Set Mining and Association Rule Induction for R.
\url{https://borgelt.net/fim4r.html}
}
