% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemsets.R
\docType{class}
\name{itemsets-class}
\alias{itemsets-class}
\alias{itemsets}
\alias{show,itemsets-method}
\alias{summary,itemsets-method}
\alias{summary.itemsets-class}
\alias{show,summary.itemsets-method}
\alias{length,itemsets-method}
\alias{nitems,itemsets-method}
\alias{labels,itemsets-method}
\alias{itemLabels,itemsets-method}
\alias{itemLabels<-,itemsets-method}
\alias{itemInfo,itemsets-method}
\alias{items,itemsets-method}
\alias{items<-,itemsets-method}
\alias{tidLists,itemsets-method}
\alias{coercion-itemsets}
\alias{coerce,itemsets,data.frame-method}
\title{Class itemsets --- A Set of Itemsets}
\usage{
itemsets(items, itemLabels = NULL, quality = data.frame())

\S4method{summary}{itemsets}(object, ...)

\S4method{length}{itemsets}(x)

\S4method{nitems}{itemsets}(x)

\S4method{labels}{itemsets}(object, ...)

\S4method{itemLabels}{itemsets}(object)

\S4method{itemLabels}{itemsets}(object) <- value

\S4method{itemInfo}{itemsets}(object)

\S4method{items}{itemsets}(x)

\S4method{items}{itemsets}(x) <- value

\S4method{tidLists}{itemsets}(x)
}
\arguments{
\item{items}{a \link{itemMatrix} or an object that can be converted using \code{\link[=encode]{encode()}}.}

\item{itemLabels}{item labels used for \code{encode()}.}

\item{quality}{a data.frame with quality information (one row per itemset).}

\item{object, x}{the object}

\item{...}{further argments}

\item{value}{replacement value}
}
\description{
The \code{itemsets} class represents a set of itemsets and the associated
quality measures.
}
\details{
Itemsets are usually created by calling an association rule mining algorithm
like \code{\link[=apriori]{apriori()}}.  Itemsets store the items as an object of class
\link{itemMatrix}.

To create itemsets manually, the itemMatrix for the items of the itemsets
can be created using \link{itemCoding}.  An example is in the Example
section below.

Mined itemsets sets typically contain several interest measures accessible
with the \code{quality()} method.  Additional measures can be
calculated via \code{\link[=interestMeasure]{interestMeasure()}}.
}
\section{Functions}{
\itemize{
\item \code{summary,itemsets-method}: create a summary

\item \code{length,itemsets-method}: get the number of itemsets.

\item \code{nitems,itemsets-method}: get the number of items (columns) in the current encoding.

\item \code{labels,itemsets-method}: get the itemset labels.

\item \code{itemLabels,itemsets-method}: get the item labels.

\item \code{itemLabels<-,itemsets-method}: replace the item labels.

\item \code{itemInfo,itemsets-method}: get item info data.frame.

\item \code{items,itemsets-method}: get items as an itemMatrix.

\item \code{items<-,itemsets-method}: with a different itemMatrix.

\item \code{tidLists,itemsets-method}: get tidLists stored in the object (if any).
}}

\section{Objects from the Class}{
 Objects are the result of calling the
functions \code{\link[=apriori]{apriori()}} (e.g., with \code{target = "frequent itemsets"} in the parameter list) or \code{\link[=eclat]{eclat()}}.  Objects can also
be created by calls of the form

\code{new("itemsets", ...)}

or by using the constructor function

\code{itemsets(items, itemLabels, quality = data.frame())}.

\code{items} need to be a list describing the items (using labels or item
ids) and \code{itemLabels} needs to be a vector of all possible item labels
(character) or a transactions object to copy the item coding (see
\link{itemCoding} for details).
}

\section{Coercion}{

\itemize{
\item as("itemsets", "data.frame")
}
}

\examples{
data("Adult")

## Mine frequent itemsets with Eclat.
fsets <- eclat(Adult, parameter = list(supp = 0.5))

## Display the 5 itemsets with the highest support.
fsets.top5 <- sort(fsets)[1:5]
inspect(fsets.top5)

## Get the itemsets as a list
as(items(fsets.top5), "list")

## Get the itemsets as a binary matrix
as(items(fsets.top5), "matrix")

## Get the itemsets as a sparse matrix, a ngCMatrix from package Matrix.
## Warning: for efficiency reasons, the ngCMatrix you get is transposed
as(items(fsets.top5), "ngCMatrix")

## Manually create itemsets with the item coding in the Adult dataset
## and calculate some interest measures
twoitemsets <- itemsets(
  items = list(
    c("age=Young", "relationship=Unmarried"),
    c("age=Old")
  ), itemLabels = Adult)

quality(twoitemsets) <- data.frame(support = interestMeasure(twoitemsets,
  measure = c("support"), transactions = Adult))

inspect(twoitemsets)
}
\seealso{
Superclass: \link{associations}

Other association functions: 
\code{\link{is.redundant}()}
}
\author{
Michael Hahsler
}
\concept{association functions}
\keyword{classes}
