% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Mushroom}
\alias{Mushroom}
\title{Mushroom Data Set}
\format{
Object of class \code{transactions}.
}
\source{
The data set was obtained from the UCI Machine Learning Repository
at \url{https://archive.ics.uci.edu/ml/datasets/Mushroom}.
}
\description{
The \code{Mushroom} data set includes descriptions of hypothetical samples
corresponding to 23 species of gilled mushrooms in the Agaricus and Lepiota
Family.  It contains information about 8124 mushrooms (transactions).  4208
(51.8\\%) are edible and 3916 (48.2\\%) are poisonous. The data contains 22
nominal features plus the class attribute (edible or not). These features
were translated into 114 items.
}
\references{
Alfred A. Knopf (1981). Mushroom records drawn from The Audubon
Society Field Guide to North American Mushrooms. G. H. Lincoff (Pres.), New
York.
}
\author{
Michael Hahsler
}
\keyword{datasets}
