% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample}
\alias{sample}
\alias{sample,itemMatrix-method}
\alias{sample,associations-method}
\title{Random Samples and Permutations}
\usage{
\S4method{sample}{itemMatrix}(x, size, replace = FALSE, prob = NULL, ...)

\S4method{sample}{associations}(x, size, replace = FALSE, prob = NULL, ...)
}
\arguments{
\item{x}{object to be sampled from (a set of \link{associations} or
\link{transactions}).}

\item{size}{sample size.}

\item{replace}{a logical. Sample with replacement?}

\item{prob}{a numeric vector of probability weights.}

\item{...}{further arguments.}
}
\value{
An object of the same class as \code{x}.
}
\description{
Provides the generic function \code{sample()} and methods to sample from \link{transactions} and
\link{associations}.
}
\examples{
data("Adult")

## sample with replacement
s <- sample(Adult, 500, replace = TRUE)
s
}
\seealso{
Other preprocessing:
\code{\link[=discretize]{discretize()}},
\code{\link{hierarchy}},
\code{\link{itemCoding}},
\code{\link[=merge]{merge()}}

Other associations functions:
\code{\link[=abbreviate]{abbreviate()}},
\code{\link{associations-class}},
\code{\link{c}},
\code{\link[=duplicated]{duplicated()}},
\code{\link{extract}},
\code{\link[=inspect]{inspect()}},
\code{\link[=is.closed]{is.closed()}},
\code{\link[=is.generator]{is.generator()}},
\code{\link[=is.maximal]{is.maximal()}},
\code{\link[=is.redundant]{is.redundant()}},
\code{\link[=is.significant]{is.significant()}},
\code{\link[=is.superset]{is.superset()}},
\code{\link{itemsets-class}},
\code{\link[=match]{match()}},
\code{\link{rules-class}},
\code{\link{sets}},
\code{\link[=size]{size()}},
\code{\link[=sort]{sort()}},
\code{\link[=unique]{unique()}}

Other itemMatrix and transactions functions:
\code{\link[=abbreviate]{abbreviate()}},
\code{\link{c}},
\code{\link[=crossTable]{crossTable()}},
\code{\link[=duplicated]{duplicated()}},
\code{\link{extract}},
\code{\link{hierarchy}},
\code{\link{image}},
\code{\link[=inspect]{inspect()}},
\code{\link[=is.superset]{is.superset()}},
\code{\link[=itemFrequency]{itemFrequency()}},
\code{\link[=itemFrequencyPlot]{itemFrequencyPlot()}},
\code{\link{itemMatrix-class}},
\code{\link[=match]{match()}},
\code{\link[=merge]{merge()}},
\code{\link[=random.transactions]{random.transactions()}},
\code{\link{sets}},
\code{\link[=size]{size()}},
\code{\link[=supportingTransactions]{supportingTransactions()}},
\code{\link{tidLists-class}},
\code{\link{transactions-class}},
\code{\link[=unique]{unique()}}
}
\author{
Michael Hahsler
}
\concept{associations functions}
\concept{itemMatrix and transactions functions}
\concept{preprocessing}
\keyword{manip}
